pyproject.toml
[tool.poetry]
name = "f-lab-python-backend-project-template"
version = "0.1.0"
description = ""
authors = ["ydrah <ydrah@f-lab.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
sqlmodel = "^0.0.14"
pydantic-settings = "^2.1.0"
aiosqlite = "^0.19.0"
greenlet = "^3.0.2"
python-jose = "^3.3.0"
email-validator = "^2.1.1"
pytest-mock = "^3.14.0"
aiomysql = "^0.2.0"
bcrypt = "^4.1.2"
pytest-asyncio = "^0.23.6"
freezegun = "^1.5.0"
starlette-csrf = "^3.0.0"
cryptography = "^42.0.7"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
prometheus-fastapi-instrumentator = "^7.0.0"
gunicorn = "^22.0.0"
python-logging-loki = "^0.3.1"


[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
isort = "^5.12.0"
pycln = "^2.3.0"
pytest = "^7.4.3"
pre-commit = "^3.5.0"
pytest-asyncio = "^0.23.2"
httpx = "^0.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
